More modularization + Lint
