+++ lib/Cache.js
@@ -3,12 +3,12 @@
 const flatCache = require('flat-cache')
 
 module.exports = class Cache {
-  constructor(name, path, cacheTime) {
+  constructor (name, path, cacheTime) {
     this.cache = flatCache.load(name, path)
     this.now = new Date().getTime()
     this.expire = this.now + cacheTime
   }
-  getKey(key) {
+  getKey (key) {
     var value = this.cache.getKey(key)
     if (value === undefined || value.expire < this.now) {
       return undefined
@@ -16,13 +16,13 @@ module.exports = class Cache {
       return value.data
     }
   }
-  setKey(key, value) {
+  setKey (key, value) {
     this.cache.setKey(key, {
       expire: this.expire,
       data: value
     })
   }
-  save() {
+  save () {
     this.cache.save(true)
   }
 }
