+++ lib/get-urls/get-urls-from-crawler.js
@@ -6,7 +6,7 @@ module.exports = (url, cacheTime) => {
       const path = require('path')
       const clc = require('cli-color')
       const urlHash = require('crypto').createHash('sha1').update(url).digest('hex')
-      const Cache = require('../cache')
+      const Cache = require('../Cache')
       const Crawler = require('simplecrawler')
 
       var greenOnBlack = clc.xterm(47).bgXterm(0)
