+++ lib/validate-pages/get-result.js
@@ -14,5 +14,4 @@ module.exports = async (url, cache, cacheTime) => {
     cache.setKey(url, data)
     return JSON.parse(data)
   }
-
 }
