+++ lib/validate-pages/is-not-found.js
@@ -1,0 +1,9 @@
+'use strict'
+
+module.exports = (result) => {
+  try {
+    return result.messages[0].type === 'non-document-error'
+  } catch (error) {
+    return false
+  }
+}
