+++ lib/validate-pages/is-valid-html.js
@@ -1,0 +1,6 @@
+'use strict'
+
+module.exports = (result, quiet) => {
+  var err = (quiet) ? result.messages.filter(m => m.type === 'error') : result.messages
+  return err.length === 0
+}
