+++ lib/validate-pages/process-pages.js
@@ -14,14 +14,16 @@ module.exports = async (pagesToValidate, options) => {
   const path = require('path')
   const pagesHash = require('crypto').createHash('sha1').update(pagesToValidate.join('_')).digest('hex')
   const clc = require('cli-color')
-  const Cache = require('../cache')
-  const getResult = require('./get-result.js')
-  const printError = require('./print-error.js')
+  const Cache = require('../Cache')
+  const getResult = require('./get-result')
+  const printError = require('./print-error')
+  const printFinish = require('./print-finish')
+  const isNotFound = require('./is-not-found')
+  const isValidHtml = require('./is-valid-html')
+  const failFast = require('./fail-fast')
 
   var greenOnBlack = clc.xterm(47).bgXterm(0)
   var redOnBlack = clc.xterm(196).bgXterm(0)
-  var yellowOnBlack = clc.xterm(11).bgXterm(0)
-  var cyanOnBlack = clc.xterm(14).bgXterm(0)
 
   var cachePath = path.resolve(`${__dirname}/../../cache/w3c/`)
   var cache = new Cache(pagesHash, cachePath, options.cacheTime)
@@ -30,30 +32,10 @@ module.exports = async (pagesToValidate, options) => {
   console.log(`\nEvaluating a total of ${pagesTotal} pages`)
 
   let pagesNotFound = []
-  const isNotFound = (result) => {
-    try {
-      return result.messages[0].type === 'non-document-error'
-    } catch (error) {
-      return false
-    }
-  }
-
   let pagesFail = []
-  const isValidHtml = (result, quiet) => {
-    var err = (quiet) ? result.messages.filter(m => m.type === 'error') : result.messages
-    return err.length === 0
-  }
-
-  console.log('__________________________________')
-
-  const failFast = (pagesNotFound, pagesFail) => {
-    if (pagesNotFound.length > 0 || pagesFail.length > 0) {
-      console.log(`\n${redOnBlack('Fast Fail Mode: Site Failed Validation')} \n`)
-      process.exit(1)
-    }
-  }
 
   asyncForEach(pagesToValidate, async (page, i) => {
+    console.log('__________________________________')
     console.log(`\nChecking page ${i + 1} of ${pagesTotal} pages`)
     try {
       let result = await getResult(page, cache, options.cacheTime)
@@ -78,23 +60,9 @@ module.exports = async (pagesToValidate, options) => {
     } catch (error) {
       console.log([redOnBlack(error), page].join(' '))
     }
-    console.log('__________________________________')
   }).then(() => {
     cache.save()
-    if (pagesFail.length === 0 && pagesNotFound.length === 0) {
-      console.log(`\n${greenOnBlack('Site Validated')} No problems were found for ${options.file || options.url}`)
-    } else {
-      if (pagesFail.length !== 0) {
-        console.log(`\n${redOnBlack('Site Failed Validation')}`)
-        console.log(`\n${pagesFail.length} out of ${pagesTotal} pages failed validation for ${options.file || options.url}`)
-        pagesFail.forEach((e) => { console.log(yellowOnBlack(e)) })
-      }
-      if (pagesNotFound.length !== 0) {
-        console.log(`\n${pagesNotFound.length} out of ${pagesTotal} pages not found for ${options.file || options.url}`)
-        pagesNotFound.forEach((e) => { console.log(yellowOnBlack(e)) })
-      }
-    }
-
-    console.log(cyanOnBlack('\nFinished Checking, have an A-1 Day!') + ' \n')
+    console.log('__________________________________')
+    printFinish((options.file || options.url), pagesTotal, pagesNotFound, pagesFail)
   })
 }
